requirements.txt
affine==2.3.0
attrs==19.3.0
certifi==2019.6.16
chardet==3.0.4
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
colorful==0.5.0
colorize==1.1.0
dj-database-url==0.5.0
Django==1.11.20
django-bootstrap-modal-forms==1.4.1
django-common-helpers==0.9.2
django-crispy-forms==1.7.2
django-cron==0.5.1
django-filter==1.1.0
django-geojson==2.12.0
django-heroku==0.3.1
django-leaflet==0.24.0
django-pandas==0.6.0
django-tables2==2.0.4
django-tables2-bootstrap4==2.0.0b5
django-widget-tweaks==1.4.3
dt==0.2
enum34==1.1.10
gunicorn==19.9.0
idna==2.8
lxml==4.3.0
numpy==1.16.1
pandas==0.24.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
PTable==0.9.2
pyparsing==2.4.6
python-dateutil==2.8.0
python-decouple==3.1
pytz==2018.9
rasterio==1.1.3
requests==2.22.0
Shapely==1.7.0
six==1.12.0
snuggs==1.4.7
tzlocal==1.5.1
urllib3==1.25.3
whitenoise==4.1.2
xlrd==1.2.0
XlsxWriter==1.1.7
